+++ src/main/java/name/divinityunbound/item/custom/WandOfCapturingItem.java
@@ -1,13 +1,7 @@
 package name.divinityunbound.item.custom;
 
-import io.github.ladysnake.pal.AbilitySource;
-import io.github.ladysnake.pal.Pal;
-import name.divinityunbound.DivinityUnbound;
-import name.divinityunbound.entity.effect.ModStatusEffect;
 import net.minecraft.client.item.TooltipContext;
 import net.minecraft.entity.*;
-import net.minecraft.entity.effect.StatusEffectInstance;
-import net.minecraft.entity.mob.MobEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
@@ -16,17 +10,13 @@ import net.minecraft.nbt.NbtCompound;
 import net.minecraft.text.Text;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.Hand;
-import net.minecraft.nbt.NbtCompound;
-import net.minecraft.util.TypedActionResult;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
-import net.minecraft.util.math.random.Random;
-import net.minecraft.world.ServerWorldAccess;
 import net.minecraft.world.World;
-import net.minecraft.world.event.GameEvent;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
+import java.util.Optional;
 
 // Done with the help of:
 // https://github.com/InnovativeOnlineIndustries/Industrial-Foregoing/tree/1.20
@@ -53,13 +43,16 @@ public class WandOfCapturingItem extends Item {
                 return false;
             }
             //player.sendMessage(Text.literal("Stack nbt " + stack.getNbt().toString()));
-            Entity entity = getEntityFromStack(stack, worldIn, false, false);
+            Optional<Entity> ent = getEntityFromStack(stack, worldIn, false, false);
+            Entity entity = ent.get();
             //player.sendMessage(Text.literal("Has entity " + (entity != null)));
             if (entity != null) {
+                //player.sendMessage(Text.literal(stack.getSubNbt("entity").toString()));
                 entity.refreshPositionAndAngles(pos.getX() + 0.5,
                         pos.getY() + 1, pos.getZ() + 0.5,
                         worldIn.getRandom().nextFloat() * 360.0F, 0.0F);
-                stack.setNbt(null);
+                //stack.setNbt(null);
+                stack.removeSubNbt("entity");
                 boolean spawned = worldIn.spawnEntity(entity);
                 //player.sendMessage(Text.literal("Spawned Entity " + spawned));
                 return spawned;
@@ -69,11 +62,13 @@ public class WandOfCapturingItem extends Item {
     }
 
     @Nullable
-    public Entity getEntityFromStack(ItemStack stack, World world, boolean withInfo, boolean applyDuplicateFilter) {
+    public Optional<Entity> getEntityFromStack(ItemStack stack, World world, boolean withInfo, boolean applyDuplicateFilter) {
         if (stack.getNbt() != null) {
-            EntityType type = EntityType.fromNbt(stack.getNbt()).orElse(null);
-            Entity entity = type.create(world);
-            return entity;
+            if (stack.getNbt().contains("entity")) {
+                //EntityType<?> type = EntityType.fromNbt(stack.getSubNbt("entity")).orElse(null);
+                //Entity entity = type.create(world);
+                return EntityType.getEntityFromNbt(stack.getSubNbt("entity"), world);
+            }
         }
         return null;
     }
@@ -83,14 +78,23 @@ public class WandOfCapturingItem extends Item {
         super.useOnEntity(stack, user, entity, hand);
 
         if (!user.getWorld().isClient) {
-            if (stack.getNbt() != null) return ActionResult.FAIL;
+            if (stack.getNbt() != null
+                    && stack.getNbt().contains("entity")) {
+                return ActionResult.FAIL;
+            }
             if(!(entity instanceof PlayerEntity)) {
+                stack.removeSubNbt("entity");
                 NbtCompound nbt = new NbtCompound();
                 entity.saveSelfNbt(nbt);
+
+                NbtCompound nbt2 = new NbtCompound();
+                nbt2.put("entity", nbt);
                 //user.sendMessage(Text.literal("Entity nbt " + nbt.get("id")));
-                stack.setNbt(nbt);
+                stack.setNbt(nbt2);
+                //stack.setNbt(nbt);
                 //user.sendMessage(Text.literal("Entity nbt " + stack.getNbt().toString()));
                 entity.remove(Entity.RemovalReason.KILLED);
+                user.setStackInHand(hand, stack);
                 return ActionResult.SUCCESS;
             }
         }
@@ -106,7 +110,12 @@ public class WandOfCapturingItem extends Item {
     public void appendTooltip(ItemStack stack, @Nullable World world, List<Text> tooltip, TooltipContext context) {
         tooltip.add(Text.translatable("tooltip.divinityunbound.wand_of_capturing"));
         if (stack.getNbt() != null) {
-            tooltip.add(Text.literal("Contains: " + stack.getNbt().get("id")));
+            if (stack.getNbt().contains("entity")) {
+                NbtCompound nbt = stack.getSubNbt("entity");
+                if (nbt != null) {
+                    tooltip.add(Text.literal("Contains: " + nbt.get("id")));
+                }
+            }
         }
         super.appendTooltip(stack, world, tooltip, context);
     }